+++ Phoenix/PHShebangPreprocessor.m
@@ -2,6 +2,7 @@
  * Phoenix is released under the MIT License. Refer to https://github.com/kasper/phoenix/blob/master/LICENSE.md
  */
 
+#import "NSTask+PHExtension.h"
 #import "PHShebangPreprocessor.h"
 
 @implementation PHShebangPreprocessor
@@ -21,34 +22,6 @@
     return command;
 }
 
-+ (NSError *) readErrorFromStandardError:(NSPipe *)standardError {
-
-    NSData *errorData = [standardError.fileHandleForReading readDataToEndOfFile];
-
-    if (errorData.length == 0) {
-        return nil;
-    }
-
-    NSString *reason = [[NSString alloc] initWithData:errorData encoding:NSUTF8StringEncoding];
-
-    return [NSError errorWithDomain:PHShebangPreprocessorErrorDomain
-                               code:PHShebangPreprocessorErrorCode
-                           userInfo:@{ NSLocalizedDescriptionKey: @"Preprocessing failed.",
-                                       NSLocalizedFailureReasonErrorKey: reason }];
-}
-
-+ (NSString *) readOutputFromStandardOutputFile:(NSFileHandle *)standardOutputFile {
-
-    NSString *output = [[NSString alloc] initWithData:[standardOutputFile readDataToEndOfFile]
-                                             encoding:NSUTF8StringEncoding];
-
-    // Remove shebang-directive if it is still present
-    NSScanner *outputScanner = [NSScanner scannerWithString:output];
-    [self scanCommand:outputScanner];
-
-    return [output substringFromIndex:outputScanner.scanLocation];
-}
-
 #pragma mark - Preprocessing
 
 + (NSString *) process:(NSString *)script atPath:(NSString *)path error:(NSError **)error {
@@ -62,33 +35,28 @@
 
     /* Launch process */
 
-    NSTask *task = [[NSTask alloc] init];
-    NSPipe *standardOutput = [NSPipe pipe];
-    NSPipe *standardError = [NSPipe pipe];
-    NSFileHandle *standardOutputFile = standardOutput.fileHandleForReading;
-
-    task.launchPath = [NSProcessInfo processInfo].environment[@"SHELL"];
-    task.standardOutput = standardOutput;
-    task.standardError = standardError;
-    task.arguments = @[ @"-l", @"-c", [NSString stringWithFormat:@"%@ %@", command, path] ];
-
-    [task launch];
-    [task waitUntilExit];
-
-    /* Read output */
-
-    NSError *taskError = [self readErrorFromStandardError:standardError];
-
+    NSError *taskError;
+    NSString *output = [NSTask outputFromLaunchedTaskWithLaunchPath:@"/bin/bash"
+                                                        environment:@{ @"PATH": [NSTask userPath] }
+                                                          arguments:@[ @"-c", [NSString stringWithFormat:@"%@ %@", command, path] ]
+                                                              error:&taskError];
     if (taskError) {
 
         if (error) {
-            *error = taskError;
+            *error = [NSError errorWithDomain:PHShebangPreprocessorErrorDomain
+                                         code:PHShebangPreprocessorErrorCode
+                                     userInfo:@{ NSLocalizedDescriptionKey: @"Preprocessing failed.",
+                                                 NSUnderlyingErrorKey: taskError }];
         }
 
         return script;
     }
 
-    return [self readOutputFromStandardOutputFile:standardOutputFile];
+    // Remove shebang-directive if it is still present
+    NSScanner *outputScanner = [NSScanner scannerWithString:output];
+    [self scanCommand:outputScanner];
+
+    return [output substringFromIndex:outputScanner.scanLocation];
 }
 
 @end